databaes-backend/requirements.txt
astroid==2.3.3
Django==2.2.7
django-filter==2.2.0
djangorestframework==3.10.3
isort==4.3.21
lazy-object-proxy==1.4.3
Markdown==3.1.1
mccabe==0.6.1
postgres==3.0.0
psycopg2==2.8.4
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pylint==2.4.3
pytz==2019.3
six==1.13.0
sqlparse==0.3.0
typed-ast==1.4.0
wrapt==1.11.2
